requirements.txt
requests
argparse
giturlparse
gitpython
PyGithub

pandas
keras
sklearn
tensorflow
titlecase
numpy